+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 * Define `Adj.Algebra.Unit` module
 * Define `Adj.Program.Instruction` module
 * Define `Adj.Program.Construction` module
+* Define `:.` method in `Categroy` typeclass